Splitting up colorbar into BorderVisual, rewriting ColorBar to work with transforms correctly
