+++ vispy/visuals/__init__.py
@@ -37,4 +37,5 @@ from .tube import TubeVisual  # noqa
 from .visual import BaseVisual, Visual, CompoundVisual  # noqa
 from .volume import VolumeVisual  # noqa
 from .xyz_axis import XYZAxisVisual  # noqa
+from .border import BorderVisual  # noqa
 from .colorbar import ColorBarVisual  # noqa
